Integrated fix to all mock servers
