+++ src/behaving/mail/mock.py
@@ -9,6 +9,21 @@ import asyncore
 output_dir = None
 
 
+def getUniqueFilename(recipient_dir, ext='tmp'):
+    filename = (
+        time.strftime("%Y-%m-%d-%H%M%S", time.gmtime(time.time())))
+    dest = os.path.join(recipient_dir, "%s.%s" % (filename, ext))
+    i = 0
+    while os.path.isfile(dest):
+        i += 1
+        if i > 1000:
+            raise IOError("Tried too many filenames like: %s" % dest)
+        filename = filename + "_" + str(i)
+        dest = os.path.join(recipient_dir, "%s.%s" % (filename, ext))
+
+    return dest
+
+
 class DebuggingServer(smtpd.DebuggingServer):
     def __init__(self, localaddr, remoteaddr, log_to_stdout=True):
         global output_dir
@@ -26,18 +41,11 @@ class DebuggingServer(smtpd.DebuggingServer):
             sys.stdout.flush()
         if self.path is None:
             return
-        filename = time.strftime("%Y-%m-%d-%H%M%S", time.gmtime(time.time()))
         for addr in rcpttos:
             path = os.path.join(self.path, addr)
             if not os.path.exists(path):
                 os.makedirs(path)
-            dest = os.path.join(path, "%s.eml" % filename)
-            index = 1
-            while os.path.exists(dest):
-                if index > 1000:
-                    raise IOError("Tried too many filenames like: %s" % dest)
-                dest = os.path.join(path, "%s-%s.eml" % (filename, index))
-                index = index + 1
+            dest = getUniqueFilename(path, "eml")
             with open(dest, "w") as f:
                 f.write(data)
